Harden code quality boundaries
