+++ desktop/runtime/composer-state.cts
@@ -1,5 +1,5 @@
 import type { ThinkingLevel } from "@mariozechner/pi-agent-core";
-import { type Model, supportsXhigh } from "@mariozechner/pi-ai";
+import { supportsXhigh } from "@mariozechner/pi-ai";
 import type { AgentSession } from "@mariozechner/pi-coding-agent";
 import type {
   ComposerModel,
