+++ src/app/app-shell/AppShellLayout.tsx
@@ -27,7 +27,6 @@ export function AppShellLayout({ controller }: AppShellLayoutProps) {
   const {
     activeComposerState,
     activeThreadData,
-    archivedThreads,
     collapsedProjectIds,
     composerProjectId,
     currentProjectName,
