+++ src/app/views/landing-overview-content.ts
@@ -1,7 +1,3 @@
-type LandingOverviewSection = {
-  markdown: string;
-};
-
 type LandingOverviewContent = {
   title: string;
   roadmap: {
